Клиентская часть rest взаимодействия по оплате товаров. Хранение товаров в кеше Redis Тесты.

+++ payment-service/src/main/java/ru/yandex/practicum/payment/model/User.java
@@ -16,5 +16,5 @@ import java.math.BigDecimal;
 public class User {
     @Id
     private String userId;
-    private BigDecimal saldo;
+    private Long saldo;
 }
