+++ payment-service/src/main/java/ru/yandex/practicum/payment/util/BalanceGenerator.java
@@ -7,8 +7,8 @@ public class BalanceGenerator {
 
     private static final Random random = new Random();
 
-    public static BigDecimal get() {
-        double mean =25.0;    // среднее: 25 000 руб.
+    public static Long get() {
+        double mean = 25.0;    // среднее: 25 000 руб.
         double stdDev = 15.0; // стандартное отклонение
 
         double balance;
@@ -16,6 +16,6 @@ public class BalanceGenerator {
             balance = mean + random.nextGaussian() * stdDev;
         } while (balance < 0.0); // исключаем отрицательные значения
 
-        return BigDecimal.valueOf(Math.round(balance * 100) / 100); // округление до копеек
+        return Math.round(balance * 100) / 100; // округление до копеек
     }
 }
